removed ubuntu, updated verify tests
