+++ .github/workflows/ci.yml
@@ -31,8 +31,6 @@ jobs:
         distro:
           - centos7
           - centos8
-          - ubuntu1804
-          - ubuntu2004
 
     steps:
       - name: Check out the Repository
